Split status breakdown tests by category and generalize category metrics
